+++ game/physics/Physics_Parametric.cpp
@@ -1060,8 +1060,8 @@ idPhysics_Parametric::WriteToSnapshot
 ================
 */
 void idPhysics_Parametric::WriteToSnapshot( idBitMsgDelta &msg ) const {
-	msg.WriteInt( current.time );
-	msg.WriteInt( current.atRest );
+	msg.WriteLong( current.time );
+	msg.WriteLong( current.atRest );
 	msg.WriteFloat( current.origin[0] );
 	msg.WriteFloat( current.origin[1] );
 	msg.WriteFloat( current.origin[2] );
@@ -1135,8 +1135,8 @@ void idPhysics_Parametric::ReadFromSnapshot( const idBitMsgDelta &msg ) {
 	idVec3 linearStartValue, linearSpeed, linearBaseSpeed, startPos, endPos;
 	idAngles angularStartValue, angularSpeed, angularBaseSpeed, startAng, endAng;
 
-	current.time = msg.ReadInt();
-	current.atRest = msg.ReadInt();
+	current.time = msg.ReadLong();
+	current.atRest = msg.ReadLong();
 	current.origin[0] = msg.ReadFloat();
 	current.origin[1] = msg.ReadFloat();
 	current.origin[2] = msg.ReadFloat();
